Getting Started (all): update how to send first event
All platforms:

Not all platforms suggest "breaking your application", so move that
sentence to the platforms where it makes sense.

Go-specific:

Sending a message is slightly simpler than an error.

Also add error handling and defer call to Flush -- this is a better
pattern that doesn't induce people to think they need to call Flush
after every call to Capture*.