sdk-dev: Document sending of default integrations
Some SDKs today already send a complete list of enabled SDKs, while the
docs suggested otherwise.

Updating the suggestion to have SDKs always send a full list, because
otherwise it is harder to provide insights in Sentry as we'd need to
maintain a mapping of default integrations for each SDK for each
released version.