Edited "environment name can't contain" sentence
When I tried to name an environment with my host URL that contained protocol, hostname, and port, there wasn't created an environment in the same way as when the name contained space character. Sentry just ignored the invalid env name. 
So I believe the URL format is not available and should be specified in the documentation as well.