feat: Template variables in markdown
This primarily adds support for template variables (or otherwise javascript functions) executable within any markdown content. An example is found in the ``README.md``.

Using the system it implements the PackageRegistry and adds it to the scope as ``packages``, thus making them callable with e.g. ``packages.version()``.

Additionally this makes a few small changes to other various Gatsby configurations to improve performance.