Clarify WPF initialization instructions
Add a note to clarify that the SentrySDK should be initialized in the application constructor.

This is valuable to specify explicitly because it is not common to have a constructor in a WPF application class. App initialization is usually all done in the OnStartup() event, but attempting to initialize the Sentry SDK in the OnStartup() event leads to obscure issues with Hub and Scope stack initialization