Update basics.mdx
This is my first contribution.

I had some trouble finding out how to set a custom header for an authorization token. I tried:

```js
Sentry.init({
  dsn: "___PUBLIC_DSN___",
  headers: {
      Authorization: `Bearer ${token}`,
    },
});
```
But that was not working.

Reading the documentation it seemed that 'transportation' can be used. However, in de code the following methods are available:

[transport](https://github.com/getsentry/sentry-javascript/blob/b7baedd94c960aa05db2f5bc8318335e18965ea1/packages/core/src/transports/noop.ts)
- sendEvent
- close

Instead, the transport options should be used:

[transportOptions (typings)](https://github.com/getsentry/sentry-javascript/blob/b7baedd94c960aa05db2f5bc8318335e18965ea1/packages/types/src/transport.ts#L26)
- dsn
- headers
- httpProxy
- httpsProxy
- caCerts
- fetchParameters

What still would be a nice addition is a way for a user reading the documentation to find out what kind of options are available.