fix: Remove redirects
It does not make sense to have the same URL in the 'redirect_from' list
of multiple pages, because visiting the given URL can only redirect to
one location, not multiple.

AFAIK the has never been content in those URLs we are redirecting from,
so either we need different redirects or none (the content under the
"common" directory is used to generate pages, but "common" is not part
of the resulting URLs).

Updates 84d8a9375bfa997632fe792655a6a1bc004c2a46