Remove border and drop shadow from platform icons in guide
The platform icons from the 'platformicons' package do not have any
border or drop shadow, but type.scss adds them for any image inside
the 'prose' class, which the guide icons are automatically part of,
since they art part of the main content block.

A similar workaround is already added in platformGrid.tsx, while the
icons in the platform navigation dropdown do not need it as they are
not part of the main content (prose section).