Align notes about using standalone crash reporting backends
Use same formatting and similar language.

The references to the Apple and Electron SDKs in the minidump docs
have been removed, as the link points to the root of the native
platform docs, which in turn links to all the guides, and alternate
SDKs. This gives a more consistent flow than duplicating the content
at multiple levels (resulting in slight differences over time due to
bitrot).