Fix redirect from platform docs to specific guides
The note would end up in in guides as well due to the missing
noGuides attribute.

The PlaformSection was also excluding more than the guide redirection,
causing sub-guides to lose content. If we want to do this the change
should probably happen along with a followup to add it back to the
sub-guides unless they explicitly don't want that content.

The redirect is also now a green note, instead of a warning. Warnings
should be reserved for very extraordinary cases.