Update identify-user.mdx
Removed the email from the examples, since due to GDPR this information is deemed sensitive. Did so based on https://docs.sentry.io/platforms/javascript/data-management/sensitive-data/

Is username also considered sensitive, since it was generated by the user and can be seen as:
> Confidential IP (Intellectual Property), such as your favorite color, or your upcoming plans for world domination.
?

I think uid and ip should be fine.