fix(nextjs): Indicate that the serverless target is not supported (#3775)
`@sentry/nextjs` does not support the `serverless` target. This patch
adds a note to the manual configuration section of the nextjs docs
that surfaces this. It also recommends users use the
`experimental-serverless-trace` option as an alternative.

The note was put in an alert under manual configuration as only a few
users will be affected by this.