fix(nextjs): Update code sample in Advanced Usage of Webpack (#3677)
The advanced usage of Webpack was valid in a prior version of the SDK. This PR updates that code sample, besides removing some comments to reduce noise. If a user wants to read that noise or just learn more about it, a link to the `manual-setup` page has been included.