docs(ember): Update Ember docs for new config (#3710)
* docs(ember): Update Ember docs for new config

Now that sdk 6.5+ is out, we can use Sentry.init and pass config options similar to how the other frameworks work. This will split the config for Sentry itself from Ember specific config. The Ember config will still be set in the config files, but the Sentry options can now be sent when init is called. The only exception is that browser tracing options need to be passed as a special option (under browserTracingOptions) since Ember provides a custom integration which currently needs to be setup after the app is initialized.