Only show tracing sampling config for supported platforms
This removes a section from the docs of platforms that do not support
tracing yet, hopefully avoiding confusion.

Unfortunately, it is clear that as new platforms are supported, adding
docs is a complex process requiring similar changes in multiple places
to update lists of supported platforms.

The list in https://docs.sentry.io/product/performance/getting-started/
is maintained manually, and so is the new list added here.