Remove invalid suggestion for PHP SDK
The PHP SDK validates that IP addresses conform to a regular expression
and does not accept the string "{{auto}}".

This option doesn't make sense for PHP applications as it would infer
and use the IP address of the PHP server for all events, instead of the
IP address of the end user, which is the intended use of the ip_address
field.