fix(gha): Move license update to Craft releases
We were still relying on the dangerous `GH_SENTRY_BOT_PAT` token to update our BSL license change date. This token belongs to an account that should be strictly restricted so switching the license update mechanism to regular Craft deploys. This also has the benefit of versioning the docs repo.