Update Next.js `sentry-cli` configuration (#4105)
All the properties in a Next.js project used to be in `sentry.properties`, used by Sentry CLI. The auth token is the only property that shouldn't be committed to the VCS, so it has been moved to `.sentryclirc`. This allows committing `sentry.properties` and ignoring `.sentryclirc` to safely commit everything but the auth token.