feat(perf): Document span operations spec (#452)
This PR adds documentation around the span operations spec that we use. Span operations are a short code identifying the type of operation the span is measuring. Span operations are low cardinality attributes - they should be as general as possible while still being human readable and useful. They should avoid including high cardinality data like IDs and URLs.

For the first part of the spec, we add a table of commonly used span operations as well as some background. In a follow up patch, we will formally document the structure of span operations as part of the spec.

Co-authored-by: Rodolfo Carvalho <rodolfo.carvalho@sentry.io>