update(perf): Update performance landing and widget descriptions
This mostly updates the docs to include widget descriptions for all the actionable widgets, since most of the mini-chart / simple visualizations are covered in metrics. The new page isn't strictly necessary, I've just added it for somewhere to put these definitions for now, since we may want to add more screenshots or explanations here in the future.

Finally, made a couple changes to the main performance landing docs.