feat(nextjs): Document `hideSourceMaps` option (#4627)
This documents the option added in https://github.com/getsentry/sentry-javascript/pull/4436, to allow users to use `hidden-source-map` as their webpack `devtool` for client-side builds. (With this option enabled, the same sourcemaps are generated, but the final bundles don't have a `sourceMappingURL` comment, effectively "hiding" the sourcemaps from the browser. Doing this prevents the browser from complaining in situations where the maps are not publicly hosted.)