chore(shortcut): Add redirect from sentry app details page (#4633)
This PR will fix the broken link for Documentation on the Shortcut sentry app. The redirect to the page was never added after the slug was changed from clubhouse to shortcut. This should work (even though it's untested) since the code that creates these links is shared for all sentry apps, and is structured the same as Linear and Rookout's links.