Improve clarity for Local Scopes documentation
I stumbled across the Local Scopes documentation while looking for information on the Kotlin extensions and thought I could make it a bit more clear. I also made mentioning that exceptions and errors thrown within `with-scope` an alert as I found that to be very surprising and unexpected, particularly as that seems like different behavior from `configure-scope`.