fix(browser): Mention release health in loader limitations section (#4775)
In our current implementation of the feature, a session and a pageload transaction are essentially one and the same, which means that release health doesn't work any better than tracing does when you're using the loader. This clarifies that in the loader's Limitations section.

Fixes https://github.com/getsentry/sentry-docs/issues/4722.