fix(angular): change NPM link from `@sentry/browser` to `@sentry/angular` (#4782)
Change the NPM package link in the right nav component in the platform SDK
details panel. Instead of linking to the generic `@sentry/browser` SDK package,
it now links to the Sentry Angular SDK package (`@sentry/angular`).

Add the additional `description` property to the `config.yml` file. All changes
are based on the `@sentry/react` config.