fix(angular): mention `TraceDirective` bug and workaround in Angular SDK (#4780)
* fix(angular): mention `TraceDirective` bug and workaround in Angular SDK

Add a warning alert to `tracehelpers.mdx` in the Angular SDK guide to make users
aware that the usage of `TraceDirective` will most likely cause a compiler error
on their end. The alert links to the original issue and our intended fix which
will ship with the v7 major of the JS/Angular SDK.

Fix suboptimal usage instruction of `TraceDirective`: Instead of declaring
`TraceDirective` in the users' Angular module(s), `TraceModule` (which wraps the
directive) should in fact be imported in the module(s). This conforms to the way
how Angular component libraries should be used.

* Update src/platforms/javascript/guides/angular/components/tracehelpers.mdx

Co-authored-by: Isabel <76437239+imatwawana@users.noreply.github.com>

* Update src/platforms/javascript/guides/angular/components/tracehelpers.mdx

Co-authored-by: Isabel <76437239+imatwawana@users.noreply.github.com>

* Update src/platforms/javascript/guides/angular/components/tracehelpers.mdx

Co-authored-by: Isabel <76437239+imatwawana@users.noreply.github.com>

Co-authored-by: Isabel <76437239+imatwawana@users.noreply.github.com>