celeryproject.org compromised; use celeryq.dev
The Celery team [has not had full control][1] of their celeryproject.org
domain for well over a year. The domain is now [compromised][2] and
redirects requests to a third party.

The Celery team has [acquired a new domain][3], [docs.celeryq.dev][4].
Usage of the old domain should be urgently discontinued.

[1]: https://github.com/celery/celeryproject/issues/44#issuecomment-781323756
[2]: https://github.com/celery/celeryproject/issues/52
[3]: https://github.com/celery/celeryproject/issues/52#issuecomment-1072945869
[4]: https://docs.celeryq.dev