Update Chrome docs broken link + feature name
The docs use the term `framework blackboxing` which was changed to `Framework Ignore List`.

The URL to the `blackboxing` docs was broken and led to the main page of the Chrome Developers Docs. I searched for a Docs article about `Framework Ignore List` and only found a stub that seemed to focus more on ignoring Chrome extension scripts. I used this URL as it's the closest to the original `blackboxing` URL. I tried this fix and it worked.