chore(integrations): Update Vercel integration docs + screenshots (#5277)
This PR will update the docs for installing Vercel on Sentry SaaS. There was some outdated information and screenshots. To be safe, I left in the troubleshooting section with another older looking screenshot, but I never encountered that Issue with an ad blocker, so it might be safe to remove.