fix(native): clean up install/build instructions (#5283)
We can remove the CMAKE_BUILD_TYPE entirely for the purpose of this
documentation, because it uses our default build-type. In addition to
that it would have to be passed as '-DCMAKE_BUILD_TYPE=RelWithDebInfo`
and not via `--config` at the build-generator stage. This leads to an
error in CMake starting with version 3.20:

https://cmake.org/cmake/help/latest/release/3.20.html#other-changes

In the actual compile/build stage it could be passed via `--config` for
build-generators that provide all build-types under a single hierarchy
(like Visual Studio solutions). This is also not applicable for a
macOS example, but was silently ignored. I removed it too.

Last but not least, I got rid of the explicit '-S .' (=source) argument,
because the default value is also the cwd:

https://cmake.org/cmake/help/latest/manual/cmake.1.html#generate-a-project-buildsystem