doc: don't have to search the docs for log levels
The log level is duplicated in a number of places around the docs, for example in the JS docs here: https://docs.sentry.io/platforms/javascript/usage/set-level/

So I've just copy / pasted it in here. It'd be better to have one page with all the levels then each language can just reference that place.