fix(react): Update React Router v4/v5 to make paramaterization more clear (#5471)
To make sure we prevent possible mis-configurations with the React Router v4/v5 integration, this PR updates our docs around it. We first move the usage of `withSentryRouting` higher up, and encourage it as the default approach to use for paramaterization. We then move down the route config option below that, and offer it as a way for users to define paths outside of the usage of `withSentryRouting`.

Co-authored-by: Isabel <76437239+imatwawana@users.noreply.github.com>