Add callout on pii rule precedence (#5573)
In Relay, data scrubbing with the new config format (i.e. Advanced Data
Scrubbing) runs before applying data scrubbing settings in the old
format (Safe Fields, etc.). This means that if an advanced rule scrubs a
field, it is not recovered even if it is declared as a Safe Field.

https://github.com/getsentry/relay/blob/4894c1d3bab6111585c29c9ea3b55e1511faca60/relay-server/src/actors/processor.rs#L1830-L1843