ref: Clarify JS `tracingOrigins` Default Values (#5593)
Clarify the meaning of the JS SDK's default values for the `tracingOrigins` option of the `BrowserTracing` integration. As pointed out in https://github.com/getsentry/sentry-javascript/issues/5868 ([comment](https://github.com/getsentry/sentry-javascript/issues/5868#issuecomment-1266582799)), it is not obvious on first glance, to which requests tracing headers are attached if the default values are used. This PR therefore adds an explanation and slightly changes the ordering of the `tracingOrigins` description. 

Co-authored-by: Isabel <76437239+imatwawana@users.noreply.github.com>