ref(svelte): Update component tracking docs with new `withSentryConfig` approach (#5649)
This patch updates the Svelte component tracking docs with usage instructions for the new approach we recently introduced to the Svelte SDK. Instead of manually adding a preprocessor to their svelte config, users can now simply wrap their config with a Sentry wrapper function that takes care of adding this preprocessor for them. This new approach allows us to make more changes to the config in the future without asking our users every time to change something in their config.