Fix misleading description of `config.logger` default. (#5675)
This document is transcluded in the docs for the `sentry-rails` gem as well as the more generic `sentry-ruby` gem. This results in the statement about the default setting for `config.logger` being misleading in the Rails case, because the reader might reasonably assume that it's safe to mutate the default logger.

For example, based on these docs I might reasonably expect to be able to set `config.logger.level = Sentry::Logger::WARN` when initialising Sentry, but if I'm in a Rails app I'd be shooting myself in the foot by unwittingly mutating `Rails.logger` through a shared reference.