add `RequestData` integration (#5674)
This adds documentation for the new `RequestData` integration in the node SDK, which adds request data to events. For frameworks where we don't have wrappers in the SDK, and merely tell people how to use sentry in the docs, this also adds storing the request in scope metadata to our boilerplate.