feat(nextjs): Add `excludeServerRoutes` option to manual setup page (#5789)
This adds the new `excludeServerRoutes` option in the nextjs SDK (added in https://github.com/getsentry/sentry-javascript/pull/6207) to the docs. It also pulls all of the auto-instrumentation config into its own section in the manual setup page.

(Note: The only part of this which is actually new content is the `Opt Out of Auto-instrumentation on Specific Routes` section. Everything else is just stuff moving around.)