fix(gatsby): Fix performance sample rate option in getting started snippets (#5812)
In all of our Getting Started snippets showing people the basic SDK setup, we include `tracesSampleRate: 1.0`, so that people will send transactions by default. In the gatsby docs, though, the option name is wrong - we're turning on error sampling, not transaction sampling. This fixes it to be the right option.