fix(nextjs): Improve explanation of webpack plugin function (#5816)
Currently, in our nextjs SDK docs about the Sentry webpack plugin, we imply that all it does is upload sourcemaps. In fact, the plugin is also what causes sourcemaps to be generated in the first place, but we don't tell people that. Nor do we tell them that it automatically sets the SDK's release value. As a result, it seems from our docs like disabling the plugin only disables upload, when in fact that's not the case.

This clarifies the full scope of the plugin's function, and adds a note telling people who disable the plugin that they'll have to set the release value themselves.

Fixes https://github.com/getsentry/sentry-javascript/issues/6240.