feat(js): Add `beforeSendTransaction` option (#5814)
This adds the new `beforeSendTransaction` option to the docs, both on the options page and alongside `beforeSend` in many places in the text. Because you can't selectively turn certain phrases on and off for different platforms, I tried to indicate where I could that it's not universally supported. (Right now, I believe JS is in fact the only platform to support it, though I expect that to change in the near future.) It also does a tiny bit of wordsmithing of the text around the additions, courtesy of @lizokm and @imatwawana.

Note that the one spot where updates haven't been done in this PR is the text of the Filtering page, as it's being worked on separately in https://github.com/getsentry/sentry-docs/pull/5804. Once the changes there are done, https://github.com/getsentry/sentry-docs/pull/5838 will add `beforeSendTransaction`.