ref(replay): Update Replay installation with SDK-package import (#5971)
This PR updates the Session Replay docs after we released version 7.27.0 of the JS SDKs. With this new version, users don't need to explicitly install `@sentry/replay` anymore, as it is now a dependency of `@sentry/browser` which [exports the Replay integration](https://github.com/getsentry/sentry-javascript/pull/6508). 

* Update the session replay setup page
* Update the session replay onboarding wizard

ref #https://github.com/getsentry/sentry-javascript/issues/6326

Note: I'd like to keep this PR open for a few days (Monday or so) just to make sure 7.27.0 isn't breaking anything