Address overhead concerns when sending 100% tx (#6790)
We currently have inconsistent messaging in docs:

on the Performance at Scale page we encourage people to send 100% of transactions no matter what
in SDK docs we caution people about needing to lower the sample rate when going to production.
We need to acknowledge that there is the “ideal” situation, where sending all transactions is possible and this enables Performance to work best, but there are some scenarios where this just isn’t realistic. For some industries where latency is highly important or even just a large enough scale, it’s not an option. Adding a note about this for clarity.

Related to #6427