fix(js): Eager load registry data
The problem here was the registry data was being loaded lazily **inside
of a `eval` expression**. There was some logic to `await` the result of
the eval, but it's unclear how well that works in a JavaScript VM and is
likely cause of all of the intermittent 'Socket closed' errors we were
seeing.

This explicitly updates the gatsby-remark-variables plugin to first load
the data in the registries, and THEN handle the replacements using
synchronous accessor functions from the registry.