Update Source Maps (#6956)
* UglifyJS fixes

Some details specific to uglifyJS and others seem to have been lost and.a common sentry-cli guide is being used. Attempting to brings the uglify js specifics back and update with the new details

* just add back uglify

Generic include for sentry cli isn't as helpful. Probably a more clever way to do this, but will require restrcturing some of these alternative bundler/compilers and having includes for the specifics

release is optional but I worry this will greatly decrease the adoption of releases

* style(lint): Auto commit lint changes

* escalated

trying to piece together the parts from before, and grouping on main page

* style(lint): Auto commit lint changes

* style(lint): Auto commit lint changes

* style(lint): Auto commit lint changes

* style(lint): Auto commit lint changes

* style(lint): Auto commit lint changes

* remove bash

Users need to know to config some environment variables. This is a simple method, there are many options with variants they can choose on their own

* style(lint): Auto commit lint changes

* Update src/platform-includes/sourcemaps/overview/javascript.mdx

Co-authored-by: Luca Forstner <luca.forstner@sentry.io>

* style(lint): Auto commit lint changes

* Update src/platform-includes/sourcemaps/overview/javascript.mdx

Co-authored-by: Lukas Stracke <lukas.stracke@sentry.io>

---------

Co-authored-by: getsantry[bot] <66042841+getsantry[bot]@users.noreply.github.com>
Co-authored-by: Luca Forstner <luca.forstner@sentry.io>
Co-authored-by: Lukas Stracke <lukas.stracke@sentry.io>