Clarify that N+1s also apply to modifications (#7220)
A customer brought up that we detect N+1s on SQL queries that make modifications (e.g. INSERT, UPDATE, DELETE) and not just SELECT, even though our docs only reference the SELECT case. Add a note to the intro clarifying that modifications are also detected, and add an example of an INSERT N+1 with fix.

Co-authored-by: vivianyentran <20403606+vivianyentran@users.noreply.github.com>