docs(dev env): Add note about `.zshrc` to `make setup-pyenv` section (#1038)
The `pyenv_setup.sh` script which runs when you call `make setup-pyenv` in the `sentry` repo only considers the case for zsh wherein the user has their zsh config in `~/.zprofile`. For anyone who doesn't, the script therefore doesn't work. This adds a note to the `zsh` tab of the code snippet telling people to do that part of the setup manually in that case.