feat(suspect-commits): Add more information on creating code mappings (#8155)
Code mappings are more complicated than the current docs make it out to seem, so this adds more in-depth info on how to decide what values to use for the code mapping.

Importantly, this also calls out that this case does not work for platforms with package names. I will follow up with additional docs on what to do for those platforms.