feat(suspect-commits): Add docs for setting up Java code mappings (#8236)
Adds a new section to the Java Source Context docs.

This section is meant to explain the specifics around setting up code mappings for Java, which are different to most other platforms. I've added a link from the common docs on code mappings so that hopefully users on these platforms will now have some way of knowing how to set these up. In the future we can link directly to this page from the app, but the first step is documenting it at all.