feat(apple): Prewarmed app start tracing is stable
Since Cocoa 8.18.0 enablePreWarmedAppStartTracking is stable. This PR
also removes code samples on how to enable the feature before 8.0.0,
which we released over a year ago. As we only removed the
experimental note in the code docs of the SDK and didn't do any
other changes, we don't need to specify a minimum SDK version.