docs: Warn about ordering `@sentry_sdk.trace` and static/class method decorators (#8831)
This PR updates the Sentry Python SDK tracing docs to inform users that when tracing a static or class method using the @sentry_sdk.trace decorator, the decorator must be placed after the @staticmethod or @classmethod decorator.

Failing to adhere to this ordering will break the function (specifically when it is called on an instance rather than on the class), due to how static and class methods work in the Python language (see getsentry/sentry-python#2525). Unfortunately, because of how static and class methods work in Python it is probably impossible to change the @sentry_sdk.trace decorator to allow a flipped ordering, so alerting users in the documentation is likely the best solution.