Fix `supported`/`notSupported` frontmatter logic
The previous logic for whether `common` guide pages were applicable based on
`supported` and `notSupported` tags was not the same as the old site. It was a
little too strict, expecting `supported` to act as an allowlist and
`notSupported` as a separate blocklist. Instead, the logic combined the two.

Copied the logic from the old Gatsby site exactly (just renaming the method):
https://github.com/getsentry/sentry-docs/blob/405170000995bd8d98bd155f4e8d38ce4c19275a/src/gatsby/createPages/createPlatformPages.ts#L138-L161

Fixes a reported issue where some JavaScript guides were missing Automatic
Instrumentation pages.