Fix api/auth route conflict
The `api/auth` route added for the changelog conflicts with the `api/auth` docs
page that would normally be handled by the catchall `[[...path]]` route,
causing those doc pages to 404.

Since the newly added `api/auth` is used by the changelog, move it under the
`changelog` path to fix the route conflict.