Add sitemap.xml file
Generate a sitemap.xml during build time that is served at `/sitemap.xml`, and
add a reference to it in `robots.txt` so that crawlers can find it.
(See: https://developers.google.com/search/docs/crawling-indexing/sitemaps/build-sitemap#addsitemap)

The sitemap entries' `lastmod` property comes from the last modification time
of the source MDX files for each page. Note that this might not be completely
accurate, as updates to MDX fragments in `/includes` or `/platform-includes`
won't be considered.

Pages generated from our OpenAPI specs are missing `lastmod` properties since
we don't have a way to isolate the timestamp when individual pages change.