Add sitemap.xml file (#9173)
Generate a sitemap.xml during build time that is served statically at `/sitemap.xml`, and add a reference to it in `robots.txt` so that crawlers can find it. (See: https://developers.google.com/search/docs/crawling-indexing/sitemaps/build-sitemap#addsitemap)

The sitemap only includes the required `url` property. It would be nice if we could add `lastmod` but the modified time on the source MDX files is not available during the production Vercel build as far as I can tell.

Fixes #9095.