Fix sitemap paths (#9439)
* fix some broken sitemap.xml paths

`getDocsFrontMatter` returns an extra `/index` at the end of some paths, which
breaks the sitemap. This is already handled by its only other caller, in
`app/[[..path]]/page.tsx`.

* refactor: remove duplication by moving sitemap /index removal up

Currently the removal of a final `/index` path is duplicated in both the page
and sitemap components - move it up into getDocsFrontMatter, which they both
call.