feat(sdk): Add name spaces for metrics rate limits (#1205)
Adds a new rate limit component to quota limits in the
`x-sentry-rate-limits` header that communicates the namespace(s) that a
metrics rate limit should apply to. The data category of limits that
have this component is always `metric_bucket`, in other quotas this
component will not be set.

SDKs are instructed not to send metrics for a certain namespace if a
rate limit for that same namespace is active. The default namespace for
metrics (used by DDM) is "custom".

---------

Co-authored-by: Luca Forstner <luca.forstner@gmail.com>