feat(discover-split): Update Discover, Alerts, Dashboards docs
In preparation for the dataset split in discover and dashboards, we need
to update the docs to reflect this new change.

The changes typically involve removing `event.type:error` or
`event.type:transaction` from being mentioned in search fields because
after the dataset is selected, these are redundant. It also updates
screenshots and calls out the dataset selector more explicitly.