fix(docs): Fix missing python configure SDK code snippet #11413
In our SDK docs, for most of the platforms we were missing init SDK code snippet for python in Configure section (e.g. AIOHTTP).

This is because we were missing python.mdx as a platfrom in our platform-includes/getting-started-config/ directory.

The code snippet here is the same as for the default Python SDK configuration.