Replaced all uses of `description` with `name` for `start_span` (#11332)
To align our API to OpenTelementry we deprecated the `description` parameter and added an alias `name`. (Because in OpenTelementry a span has no `description` but only a `name`)
When switching to OpenTelementry in the next major of the Python SDK `description` will be removed and only `name` can be uses. So we want to make sure users discovering Sentry now are already prepared for that. 

(The `ai_track` is using internally the start_span and is giving its parameters to it. I removed the explixit parameter name, because it is not needed)

---------

Co-authored-by: Ivana Kellyer <ivana.kellyer@sentry.io>