feat: Improve `Expandable` style
For this, I extracted the `Alert` component stuff into a more generic `Callout` component, which the `Alert` extends.

The `Expandable` can now use the same basis, ensuring we have the same style etc, and just pass some more stuff into it.

For this, the `Callout` component also allows to pass `id` and `titleOnClick` attributes which we need for the Expandable.