feat: Improve readability of nested lists
Previously, any nested lists would be rendered as follows:

```
1. Main item
   1. Sub-item
      1. Sub-sub-item
      2. Another sub-sub-item
   2. Another sub-item
2. Another main item
```

Reading the above nested list is difficult, since the reader needs to keep track of the indentation in order to tell whether an item is a main item, sub-item, or sub-sub-item.

This PR changes the numbering format for sub-items and sub-sub-items to make the item types clearer. The example above would now render as:

```
1. Main item
   a. Sub-item
      i. Sub-sub-item
      ii. Another sub-sub-item
   b. Another sub-item
2. Another main item
```