Flutter as framework of Dart (#12866)
* Flutter as framework of Dart

This PR moves all the files of Flutter to be under Dart, same as JavaScript and Python and other language to framework relationships.

After this is merged. I will begin looking at first updating all the links in product and static-pages to respect these new redirect links.

Potentially blogs aswell.

But then importantly and funcitonally, I want to look at ensuring that feature which the Flutter SDK inherits from the Dart SDK and any example in the other direction, are properly show in the docs. For example Dio integration is not seen in flutter.

* fixing the platform icons

flutter icon should display for flutter on main page and side bar

* Replace all the old urls and redirects

* one more

* Adding Flutter URL Redirects

* add trailing /s

* remove wildcard

would otherwise need to update types or some other strangeness, and the config as is, is rather extensive

* fix test

should do dynamic fixes later

* switch to redirect.js

remove middleware redirects and put the wildcard impl in redirects

* [getsentry/action-github-commit] Auto commit

* revert bad merge

* fix bad link

it is redirecting as if it is in flutter context, but it's not a more explicit link hopefully resolves this

* stupid

---------

Co-authored-by: getsantry[bot] <66042841+getsantry[bot]@users.noreply.github.com>