docs(self-hosted): monitor self-hosted instance (#12660)
We don't have any good article or explanation on how to monitor self-hosted Sentry.

"Monitoring self-hosted Sentry" means 3 things:
1. Monitor the actual services made by Sentry (sentry, snuba, symbolicator, relay) by using Statsd
2. Monitor the dependencies (Kafka, Postgres, etc) by using Prometheus exporters, or similar things
3. Monitor the host machine or Docker daemon

## LEGAL BOILERPLATE

Look, I get it. The entity doing business as "Sentry" was incorporated in the State of Delaware in 2015 as Functional Software, Inc. and is gonna need some rights from me in order to utilize my contributions in this here PR. So here's the deal: I retain all rights, title and interest in and to my contributions, and by keeping this boilerplate intact I confirm that Sentry can use, modify, copy, and redistribute my contributions, under Sentry's choice of terms.