fix(develop): clarify trace context's client_sample_rate (#13174)
When reading the Trace Context docs, I assumed that `client_sample_rate` (like
all the other fields) was expected to be set by the client. Instead, this field
is generated in Relay. If a client sets this field, it will be ignored and
overwritten. Added a note to the docs for this field to clarify how it is meant
to be used.

Co-authored-by: Alex Krawiec <alex.krawiec@sentry.io>