Overhaul and expand API guidelines
The exposed interfaces in the API have unfortunately become far too varied over the years, and part of this can be blamed on our lack of intentional design behind them. It was always implicit: design it like it was before. Unfortunately implicit design doesn't scale as more people work on a system, so this change documents a set of design principles that we can rely on to future proof new APIs, as well as apply to legacy APIs down the road.

Some notable shifts from where we are today:

- All API endpoints must be designed as if its a public API.
- `PATCH` should not be used. It is redundant with `PUT` in our use cases. If you want to "replace a resource" (which is uncommon behavior), you would do it via `POST /resource/:id`.
- Responses should use noun-based nesting on the object vs being flat (e.g. `{project: {...}` vs `{...}` for a project response).
- Identifiers should accept both human-readable identifiers and numeric identifiers. This has regressed quite drastically over the years, both in terms of query params as well as route identifiers.
- Identifiers should use common names (`organization`) rather than ambiguous attributes (`organization_id`).

Refs #13461